+++ pkg/agent/collectors/logs/secret.go
@@ -24,24 +24,26 @@ func Secret(cluster *clusterv1.Cluster) *v1.Secret {
 	}
 }
 
-func (s *Service) GenerateAlloyLogsSecretData(ctx context.Context, cluster *clusterv1.Cluster) (map[string]string, error) {
-	lokiURL := fmt.Sprintf(common.LokiPushURLFormat, s.Config.Cluster.BaseDomain)
-	lokiRulerAPIURL := fmt.Sprintf(common.LokiBaseURLFormat, s.Config.Cluster.BaseDomain)
+func (s *Service) GenerateAlloyLogsSecretData(ctx context.Context, cluster *clusterv1.Cluster, loggingEnabled bool) (map[string]string, error) {
+	secrets := map[string]string{}
+	// Add Loki credentials if logging is enabled
+	if loggingEnabled {
+		lokiURL := fmt.Sprintf(common.LokiPushURLFormat, s.Config.Cluster.BaseDomain)
+		lokiRulerAPIURL := fmt.Sprintf(common.LokiBaseURLFormat, s.Config.Cluster.BaseDomain)
 
-	// Get Loki password
-	logsPassword, err := s.LogsAuthManager.GetClusterPassword(ctx, cluster)
-	if err != nil {
-		return nil, fmt.Errorf("failed to get loki auth password for cluster %s: %w", cluster.Name, err)
-	}
+		// Get Loki password
+		logsPassword, err := s.LogsAuthManager.GetClusterPassword(ctx, cluster)
+		if err != nil {
+			return nil, fmt.Errorf("failed to get loki auth password for cluster %s: %w", cluster.Name, err)
+		}
 
-	// Build secret environment variables map
-	secrets := map[string]string{
-		common.LokiURLKey:         lokiURL,
-		common.LokiTenantIDKey:    organization.GiantSwarmDefaultTenant,
-		common.LokiUsernameKey:    cluster.Name,
-		common.LokiPasswordKey:    logsPassword,
-		common.LokiRulerAPIURLKey: lokiRulerAPIURL,
-	}
+		// Build secret environment variables map
+		secrets[common.LokiURLKey] = lokiURL
+		secrets[common.LokiTenantIDKey] = organization.GiantSwarmDefaultTenant
+		secrets[common.LokiUsernameKey] = cluster.Name
+		secrets[common.LokiPasswordKey] = logsPassword
+		secrets[common.LokiRulerAPIURLKey] = lokiRulerAPIURL
+  }
 
 	return secrets, nil
 }